feat: move prompts to ui layer
